style: isolate the behavior of header links in a hook and navigate to each section when its link is clicked on
