+++ src/hooks/useActiveSection.tsx
@@ -1,0 +1,39 @@
+import { useEffect, useState } from 'react'
+
+/**
+ * Gets the section that is currently in view
+ * @returns {string} The active section
+ */
+export default function useActiveSection(): string {
+  const [activeSection, setActiveSection] = useState('')
+
+  useEffect(() => {
+    const sections = document.querySelectorAll('.section')
+
+    const options = {
+      root: null,
+      rootMargin: '0px',
+      threshold: 0.5,
+    }
+
+    const observer = new IntersectionObserver(entries => {
+      entries.forEach(entry => {
+        if (entry.isIntersecting) {
+          setActiveSection(entry.target.id)
+        }
+      })
+    }, options)
+
+    sections.forEach(section => {
+      observer.observe(section)
+    })
+
+    return () => {
+      sections.forEach(section => {
+        observer.unobserve(section)
+      })
+    }
+  }, [])
+
+  return activeSection
+}
